project2

+++ animal.cpp
@@ -11,7 +11,7 @@
 
 Animal::Animal() {};
 
-Animal::Animal(int age, double animalCost, int babies, int foodCost, double payOff) :
+Animal::Animal(int age, double animalCost, int babies, double foodCost, double payOff) :
     age{age},
     animalCost{animalCost},
     babyPotential{babies},
